Fix a regression caused by py3 porting that is preventing calibre from starting up on windows machines with non UTF-8 locales. Fixes #1836360 [no RUN ver 3.45.0 or 3.45.1](https://bugs.launchpad.net/calibre/+bug/1836360)
os.path.expanduser is broken in python2. When passed a unicode object it
concatenates it to a bytestring, which will lead to UnicodeDecodeError
if the bytestring happens to not be in the default encoding, ususally
UTF-8
